# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Code Red Backend API
  version: 1.3.0
  description: >
    Production-style backend contract for the Code Red coaching institute platform.
    Canonical routes are versioned under `/api/v1`. This revision aligns the contract
    with the current frontend implementation, including authority-based navigation,
    inventory requests, payroll views, promotion flows, and faculty material publishing.
    Access control follows the frontend role families: student, faculty, parent,
    and administration (`admin`, `administration`, `director`, `superadmin`). Request
    payloads are validated with Pydantic, some background work is dispatched through
    Celery using Redis, and throttled endpoints may return HTTP 429.
servers:
  - url: http://localhost:3500
    description: Local Flask API server
  - url: http://localhost:3000
    description: Frontend dev server proxy, if configured
...
paths:
...
  /api/v1/ai/generate-questions:
    post:
      tags: [AI]
      summary: Generate assessment questions
      description: Rate limited endpoint intended for faculty and administration users.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateQuestionsRequest'
      responses:
        '200':
          description: Generated questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
  /api/v1/ai/modify-questions:
    post:
      tags: [AI]
      summary: Modify generated questions
      description: Rate limited endpoint intended for faculty and administration users.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyQuestionsRequest'
      responses:
        '200':
          description: Modified questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
  /api/v1/assessments:
    get:
      tags: [Assessments]
      summary: List assessments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: classId
          schema: { type: integer }
        - in: query
          name: subjectId
          schema: { type: integer }
        - in: query
          name: published
          schema: { type: boolean }
      responses:
        '200':
          description: Assessment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assessment'
    post:
      tags: [Assessments]
      summary: Create assessment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentWrite'
      responses:
        '201':
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
  /api/v1/assessments/{assessmentId}:
    get:
      tags: [Assessments]
      summary: Get assessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      responses:
        '200':
          description: Assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
    patch:
      tags: [Assessments]
      summary: Update assessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentPatch'
      responses:
        '200':
          description: Assessment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
    delete:
      tags: [Assessments]
      summary: Delete assessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      responses:
        '204':
          description: Assessment deleted
  /api/v1/assessments/{assessmentId}/publish:
    post:
      tags: [Assessments]
      summary: Publish assessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      responses:
        '200':
          description: Assessment published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
  /api/v1/assignments:
    get:
      tags: [Assignments]
      summary: List assignments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: subjectId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Assignment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
  /api/v1/assignments/{assignmentId}/submit:
    post:
      tags: [Assignments]
      summary: Submit assignment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentSubmissionRequest'
      responses:
        '201':
          description: Assignment submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentSubmissionResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
...
    AssessmentIdPath:
      in: path
      name: assessmentId
      required: true
      schema: { type: string }
    AssignmentIdPath:
      in: path
      name: assignmentId
      required: true
      schema: { type: string }
...
  responses:
...
    Error422:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error429:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
...
    Material:
      type: object
      properties:
        id: { type: string }
        subjectId: { type: string, nullable: true }
        subjectName: { type: string, nullable: true }
        chapterId: { type: string, nullable: true }
        chapterTitle: { type: string, nullable: true }
        title: { type: string }
        unit: { type: string, nullable: true }
        week: { type: string, nullable: true }
        type: { type: string }
        description: { type: string, nullable: true }
        fileName: { type: string, nullable: true }
        documentId: { type: string, nullable: true }
        documentName: { type: string, nullable: true }
        documentUrl: { type: string, nullable: true }
        externalUrl: { type: string, nullable: true }
        uploadedAt: { type: string, format: date-time, nullable: true }
        storagePath: { type: string, nullable: true }
        uploadedBy: { type: string, nullable: true }
        isFacultyUpload: { type: boolean, nullable: true }
...
    Question:
      type: object
      properties:
        id: { type: string }
        questionText: { type: string }
        questionType: { type: string }
        options:
          type: array
          items: { type: string }
          nullable: true
        correctAnswer:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
          nullable: true
        marks: { type: integer }
        difficulty: { type: string }
    Assessment:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        classId: { type: string }
        subjectId: { type: string }
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        totalMarks: { type: integer }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        dueDate: { type: string, nullable: true }
        published: { type: boolean }
        questionsDocumentId: { type: string, nullable: true }
    Assignment:
      type: object
      properties:
        id: { type: string }
        subjectId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        dueDate: { type: string }
        totalMarks: { type: integer }
        status: { type: string }
...
    GenerateQuestionsRequest:
      type: object
      required: [subjectId, questionCount, totalMarks]
      properties:
        subjectId: { type: string }
        materials:
          type: array
          items:
            $ref: '#/components/schemas/Material'
        questionCount: { type: integer }
        totalMarks: { type: integer }
        difficultyLevel: { type: string }
        questionTypes:
          type: object
          additionalProperties: { type: integer }
        customPrompt: { type: string, nullable: true }
    ModifyQuestionsRequest:
      type: object
      required: [questions, modificationPrompt]
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        modificationPrompt: { type: string }
    AssessmentWrite:
      type: object
      required: [title, classId, subjectId, totalMarks]
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        classId: { type: string }
        subjectId: { type: string }
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        totalMarks: { type: integer }
        createdBy: { type: string, nullable: true }
        dueDate: { type: string, nullable: true }
        published: { type: boolean, nullable: true }
    AssessmentPatch:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        dueDate: { type: string, nullable: true }
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        published: { type: boolean }
    AssignmentSubmissionRequest:
      type: object
      required: [submissionUrl]
      properties:
        submissionUrl: { type: string }
    AssignmentSubmissionResponse:
      type: object
      properties:
        success: { type: boolean }
        submittedAt: { type: string, format: date-time }
...